backend/requirements.txt
# ============================================================
# Sistema de Monitoreo IoT
# Backend - Dependencias Python
# ============================================================

# ============================================================
# Framework Web
# ============================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================================
# Base de Datos y ORM
# ============================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ============================================================
# Redis (Cache y Sesiones)
# ============================================================
redis==5.0.1
hiredis==2.2.3

# ============================================================
# Validación de Datos
# ============================================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================================
# Autenticación y Seguridad
# ============================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# ============================================================
# Utilities
# ============================================================
python-dotenv==1.0.0

# ============================================================
# HTTP Client (para webhooks y Telegram)
# ============================================================
httpx==0.25.2
requests==2.31.0

# ============================================================
# Testing
# ============================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # Para TestClient de FastAPI

# ============================================================
# Validación y Linting (Desarrollo)
# ============================================================
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# ============================================================
# Documentación
# ============================================================
# FastAPI ya incluye Swagger UI y ReDoc
